docs: sd-prose docs [skip chromatic] (#1349)
